Adds possibility to start joukko mob sessions on another branch (#2)
* Renames joukko file
** From .joukko.file to .joukko
* Adds possibility to rename the current joukko branch
* Fixes finish when no changes are found
* Adds logging of remote messages on git push
* Moves file and user input functions to their own files
* Adds unit tests for start, pass, take, finish, rename and joukko
* Sets joukkoFileHasBranchName to use readJoukkoFileBranch function
* Fixes invalid input on user input branch question
* Removes branch option 'Other' from joukko take
* Updates todo list in README.md

+++ src/utils/joukko.js
@@ -1,13 +1,9 @@
-const fs = require("fs")
-const readline = require("readline")
-
 const {
-  log,
-  logError,
-  logNok,
-  logOk,
-  logWarning,
-} = require("../utils/log")
+  createFileWithJson,
+  fileExists,
+  readFileContent,
+  removeFile,
+} = require("./file")
 
 const {
   abortRebase,
@@ -20,72 +16,159 @@ const {
   getCurrentBranch,
   getDefaultBranch,
   getPathForGitRoot,
+  getRemoteBranches,
   hasChanges,
   hasUntracked,
   improveKnowledgeOfRemoteBranches,
   isGitDirectory,
   pushTo,
+  renameBranch,
   updateBranchWithPullRebase,
   getStatus,
-} = require("../utils/git")
+} = require("./git")
 
-const JOUKKO_BRANCH_FILE = ".joukko.file"
+const {
+  log,
+  logError,
+  logNok,
+  logOk,
+  logWarning,
+} = require("./log")
+
+const {
+  getUserConfirmation,
+  getUserInput,
+} = require("./userInput")
+
+
+const JOUKKO_FILE = ".joukko"
+
+const abort = message => {
+  logError(message)
+  return ""
+}
 
 const addJoukkoFileUnderGit = async () => {
   const repoRoot = await getPathForGitRoot()
-  await addFile(`${repoRoot}/${JOUKKO_BRANCH_FILE}`)
+  await addFile(`${repoRoot}/${JOUKKO_FILE}`)
 }
 
-const askUserConfirmation = question => {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  })
-  return new Promise(resolve => rl.question(`${question} (yes/no): `, answer => {
-    const upperCaseAnswer = answer.toUpperCase()
-    if (["N", "NO"].includes(upperCaseAnswer)) {
-      rl.close()
-      resolve(false)
-    } else if(["Y", "YES"].includes(upperCaseAnswer)) {
-      rl.close()
-      resolve(true)
-    } else {
-      log("Please answer yes or no.")
-      rl.close()
-      askUserConfirmation(question)
-    }
-  }))
+const getBranchNameFromRemotePath = remotePath => {
+  return remotePath.split("/").pop()
 }
 
-const askUserForBranchName = async () => {
-  const currentBranch = await getCurrentBranch()
-  return await askUserInput("Name for the mob branch", currentBranch)
+const getBranchNamesFromRemotePathBranchList = remoteBranches => {
+  const remoteBranchNames = remoteBranches.all.map(remoteBranch => getBranchNameFromRemotePath(remoteBranch))
+  const remoteBranchNamesSorted = remoteBranchNames.sort()
+  return remoteBranchNamesSorted
+}
+
+const getBranchNamesForListing = (remoteBranchNames, showOptionOther) => {
+  const branchNamesWithIndexes = remoteBranchNames.map((branchName, index) => `[\x1b[32m${index}\x1b[0m]: ${branchName}`)
+  const nextIndex = branchNamesWithIndexes.length
+  let optionCount = branchNamesWithIndexes.length
+  let branchNamesForListing
+  if (showOptionOther) {
+    branchNamesForListing = 
+    [
+      ...branchNamesWithIndexes,
+      `[\x1b[32m${nextIndex}\x1b[0m]: Other...`
+    ]
+    optionCount = optionCount + 1
+  } else {
+    branchNamesForListing = branchNamesWithIndexes
+  }
+  return branchNamesForListing
+}
+
+const askBranchFromExisting = async (branchQuestion, allowDefaultAsProposed = false, showOptionOther = true) => {
+  await improveKnowledgeOfRemoteBranches()
+  const remoteBranches = await getRemoteBranches()
+  const remoteBranchNames = getBranchNamesFromRemotePathBranchList(remoteBranches)
+  const branchNamesForListing = getBranchNamesForListing(remoteBranchNames, showOptionOther)
+  log("")
+  log(`${branchQuestion}:`)
+  branchNamesForListing.map(branchListing => log(branchListing))
+  const branchNameOptionCount = branchNamesForListing.length
+  const selection = await askOption(branchNameOptionCount)
+  if (selection == branchNameOptionCount) {
+    return await askUserForBranchName(branchQuestion, allowDefaultAsProposed, showOptionOther)
+  } else {
+    return remoteBranchNames[selection]
+  }
+}
+
+const isSelectedOptionValid = (selection, optionCount) => {
+  const acceptableOptions = [...Array(optionCount).keys()]
+  const selectedOption = parseInt(selection.trim())
+  return acceptableOptions.includes(selectedOption)
+}
+
+const askOption = async optionCount => {
+  return await askUserInput("Choose correct branch number")
+    .then(async selection => {
+      const selectionValid = isSelectedOptionValid(selection, optionCount)
+      if (selectionValid) {
+        return selection
+      } else {
+        logError("Invalid selection.")
+        return await askOption(optionCount)
+      }
+    })
+}
+
+const askUserConfirmation = question => getUserConfirmation(question)
+
+const askUserForBranchName = async (branchQuestion, allowDefaultAsProposed = false) => {
+  const defaultIsCurrent = await isCurrentBranchTheDefaultBranch()
+  if (!allowDefaultAsProposed && defaultIsCurrent) {
+    return await askUserInput(branchQuestion)
+  } else {
+    const currentBranch = await getCurrentBranch()
+    return await askUserInput(branchQuestion, currentBranch)
+  }
 }
 
 const askUserInput = (question, defaultInput = "") => {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  })
   const inputQuestion = question + (defaultInput !== "" ? ` (${defaultInput}): ` : ": ")
-  return new Promise(resolve => rl.question(inputQuestion, userInput => {
-    if (userInput !== null && userInput.trim() !== "") {
-      rl.close()
-      resolve(userInput)
-    } else {
-      if (defaultInput !== "") {
-        rl.close()
-        resolve(defaultInput)
-      } else {
-        log("Invalid input.")
-        askUserInput(question)
-      }
-    }
-  }))
+  return getUserInput(inputQuestion, defaultInput)
 }
 
-const changeToJoukkoMobBranch = async (joukkoMobBranch, remoteAlreadyUpdated = false) => {
+const changeToBranch = async (branchName = "") => {
+  if (branchName === "") {
+    throw("Branch name is empty.")
+  }
   const currentBranch = await getCurrentBranch()
+  if (branchName !== currentBranch) {
+    log(`Branch '${currentBranch}' is the current branch. Checking out the branch '${branchName}'.`)
+    await checkoutBranch(branchName)
+  } else {
+    throw("Already on the branch.")
+  }
+}
+
+const changeToDefaultBranch = async () => {
+  const defaultBranch = await getDefaultBranch()
+  if (defaultBranch === "") {
+    throw("No default branch found.")
+  }
+  const currentBranch = await getCurrentBranch()
+  if (defaultBranch !== currentBranch) {
+    log(`Branch '${currentBranch}' is the current branch. Checking out the default branch '${defaultBranch}'.`)
+    await checkoutBranch(defaultBranch)
+  } else {
+    throw("Current branch is the default branch.")
+  }
+}
+
+const changeToJoukkoMobBranch = async (joukkoMobBranch = "", remoteAlreadyUpdated = false) => {
+  if (joukkoMobBranch === "") {
+    throw("No joukko mob branch given.")
+  }
+  const currentBranch = await getCurrentBranch()
+  if (joukkoMobBranch === currentBranch) {
+    throw("Current branch is the joukko mob branch.")
+  }
   log(`Current branch '${currentBranch}' is different from the mob branch '${joukkoMobBranch}'.`)
   const joukkoMobBranchExistsLocally = await checkIfBranchExistsLocally(joukkoMobBranch)
   if (joukkoMobBranchExistsLocally) {
@@ -101,34 +184,33 @@ const changeToJoukkoMobBranch = async (joukkoMobBranch, remoteAlreadyUpdated = f
     }
     const joukkoMobBranchExistsInRemote = await checkIfBranchExistsInRemote(joukkoMobBranch)
     if (joukkoMobBranchExistsInRemote) {
-      log(`Mob branch '${joukkoMobBranch}' found in origin. Checking out.`)
+      log(`Mob branch '${joukkoMobBranch}' found in remote. Checking out.`)
       await checkoutBranch(joukkoMobBranch)
     } else {
-      log(`Mob branch '${joukkoMobBranch}' doesn't exist in origin. Creating the branch.`)
+      log(`Mob branch '${joukkoMobBranch}' doesn't exist in remote. Creating the branch.`)
       await checkoutNewBranch(joukkoMobBranch)
     }
   }
 }
 
-const checkIfJoukkoHasAlreadyBeenStarted = async () => {
-  const previousCommitAddedFiles = await getPreviousCommitAddedFiles()
-  console.log("checkIfJoukkoHasAlreadyBeenStarted ~ previousCommitAddedFiles", previousCommitAddedFiles)
-  return true
-}
-
 const checkJoukkoFile = async () => {
   const joukkoFileExists = await joukkoFileFound()
   if (!joukkoFileExists) {
-    logWarning("Joukko Mob branch file '.joukko.file' not found.")
+    logWarning("Joukko file '.joukko' not found.")
     return false
   } else {
-    const joukkoFileIsEmpty = await joukkoFileEmpty()
-    if (joukkoFileIsEmpty) {
-      logWarning("Joukko Mob branch file is empty.")
+    const joukkoFileHasBranch = await joukkoFileHasBranchName()
+    if (!joukkoFileHasBranch) {
+      logWarning("Joukko file doesn't contain branch name.")
       return false
     }
   }
   return true
+}
+
+const checkoutNewLocalBranch = async (branchName) => {
+  log(`Creating and checking out local branch '${branchName}'.`)
+  await checkoutNewBranch(branchName)
 }
 
 const checkoutUpToDateBranch = async (branchName, remoteUpdated = false) => {
@@ -148,8 +230,7 @@ const checkoutUpToDateBranch = async (branchName, remoteUpdated = false) => {
         log(`Branch '${branchName}' found locally. Checking it out.`)
         await checkoutBranch(branchName)
       } else {
-        log(`Creating and checking out local branch '${branchName}'.`)
-        await checkoutNewBranch(branchName)
+        await checkoutNewLocalBranch(branchName)
       }
     }
   }
@@ -158,12 +239,13 @@ const checkoutUpToDateBranch = async (branchName, remoteUpdated = false) => {
 const createJoukkoBranchFile = async (branchName, silent = false) => {
   try {
     const repoRoot = await getPathForGitRoot()
-    fs.writeFileSync(`${repoRoot}/${JOUKKO_BRANCH_FILE}`, branchName)
+    const branchJson = { branch: branchName }
+    createFileWithJson(repoRoot, JOUKKO_FILE, JSON.stringify(branchJson))
     if (!silent) {
-      log("Mob branch file created.")
+      log("Joukko branch file created.")
     }
-  } catch(error) {
-    logWarning("Could not create mob branch file.")
+  } catch (error) {
+    logWarning("Could not create joukko branch file.")
     throw(error)
   }
 }
@@ -171,6 +253,12 @@ const createJoukkoBranchFile = async (branchName, silent = false) => {
 const isBranchCurrentBranch = async branchName => {
   const currentBranch = await getCurrentBranch()
   return branchName === currentBranch 
+}
+
+const isCurrentBranchTheDefaultBranch = async () => {
+  const currentBranch = await getCurrentBranch()
+  const defaultBranch = await getDefaultBranch()
+  return defaultBranch == currentBranch
 }
 
 const joukkoFileEmpty = async () => {
@@ -185,12 +273,17 @@ const joukkoFileEmpty = async () => {
 const joukkoFileFound = async () => {
   const repoRoot = await getPathForGitRoot()
   try {
-    if (fs.existsSync(`${repoRoot}/${JOUKKO_BRANCH_FILE}`)) {
-      return true
-    } else {
-      return false
-    }
-  } catch(err) {
+    return fileExists(repoRoot, JOUKKO_FILE)
+  } catch (error) {
+    return false
+  }
+}
+
+const joukkoFileHasBranchName = async () => {
+  try {
+    const joukkoBranch = await readJoukkoFileBranch()
+    return !!joukkoBranch
+  } catch (error) {
     return false
   }
 }
@@ -202,17 +295,40 @@ const preCheckForFinishAndPassOk = async () => {
     logNok("Pre-check")
     return false
   }
-  const joukkoFileExists = await checkJoukkoFile()
-  if (!joukkoFileExists) {
-    logWarning("Joukko branch file not found.")
+  const joukkoFileWithBranchExists = await checkJoukkoFile()
+  if (!joukkoFileWithBranchExists) {
+    logWarning("Joukko branch file with branch not found.")
     logNok("Pre-check")
     return false
   }
-  const joukkoMobBranch = await readJoukkoFileContent()
+  const joukkoMobBranch = await readJoukkoFileBranch()
   const currentBranch = await getCurrentBranch()
 
   if (joukkoMobBranch !== currentBranch) {
     logWarning(`Current branch '${currentBranch}' is different from the mob branch '${joukkoMobBranch}'.`)
+    logNok("Pre-check")
+    return false
+  }
+  logOk("Pre-check")
+  return true
+}
+
+const preCheckForRenameOk = async () => {
+  const isGit = await isGitDirectory()
+  if (!isGit) {
+    logWarning("Directory is not a git directory.")
+    logNok("Pre-check")
+    return false
+  }
+  const currentBranchIsDefaultBranch = await isCurrentBranchTheDefaultBranch()
+  if (currentBranchIsDefaultBranch) {
+    logWarning("Current branch is the default branch. Can't rename.")
+    logNok("Pre-check")
+    return false
+  }
+  const joukkoFileWithBranchExists = await checkJoukkoFile()
+  if (!joukkoFileWithBranchExists) {
+    logWarning("Joukko branch file with branch not found.")
     logNok("Pre-check")
     return false
   }
@@ -230,12 +346,6 @@ const preCheckForStartOk = async () => {
   const changes = await hasChanges()
   if (changes) {
     logWarning("Changes found.")
-    logNok("Pre-check")
-    return false
-  }
-  const joukkoFileExists = await joukkoFileFound()
-  if (joukkoFileExists) {
-    logWarning("Joukko branch file found. Mob programming session already started.")
     logNok("Pre-check")
     return false
   }
@@ -273,25 +383,44 @@ const pushToBranch = async (branch, options = []) => {
 
 const pushToMobBranch = async (options = []) => {
   const remote = "origin"
-  const mobBranch = await readJoukkoFileContent()
+  const mobBranch = await readJoukkoFileBranch()
   await pushTo(remote, mobBranch, options)
+}
+
+const readJoukkoFileBranch = async () => {
+  try {
+    const joukkoFileContent = await readJoukkoFileContent()
+    if (joukkoFileContent && joukkoFileContent.trim() != "") {
+      const joukkoFileJson = JSON.parse(joukkoFileContent)
+      const joukkoBranch = joukkoFileJson.branch
+      if (joukkoBranch && joukkoBranch.trim() != "") {
+        return joukkoFileJson.branch
+      } else {
+        throw("No branch name in joukko file.")
+      }
+    } else {
+      throw("Joukko file is empty.")
+    }
+  } catch (error) {
+    logError(error)
+    throw("Could not read branch from joukko file.")
+  }
 }
 
 const readJoukkoFileContent = async () => {
   const repoRoot = await getPathForGitRoot()
   try {
-    const fileContent = fs.readFileSync(`${repoRoot}/${JOUKKO_BRANCH_FILE}`, "utf-8")
-    return fileContent.trim()
-  } catch(err) {
-    logError(err)
+    return readFileContent(repoRoot, JOUKKO_FILE)
+  } catch (error) {
+    logError(error)
   }
 }
 
 const reCheckoutBranchFromRemote = async branchName => {
   const branchIsCurrentBranch = await isBranchCurrentBranch(branchName)
+  const defaultBranch = await getDefaultBranch()
   try {
     if (branchIsCurrentBranch) {
-      const defaultBranch = await getDefaultBranch()
       if (defaultBranch !== "" && defaultBranch !== branchName) {
         log(`Branch '${branchName}' is the current branch. Checking out the default branch '${defaultBranch}'.`)
         await checkoutBranch(defaultBranch)
@@ -300,17 +429,21 @@ const reCheckoutBranchFromRemote = async branchName => {
       }
     }
     if (branchIsCurrentBranch || await checkIfBranchExistsLocally(branchName)) {
-      try {
-        log(`Removing local branch '${branchName}'.`)
-        await deleteBranch(branchName)
-      } catch(deleteError) {
-        logError(`Could not remove local branch '${branchName}'.`)
-        throw(deleteError)
+      if (defaultBranch !== "" && defaultBranch === branchName) {
+        throw("The given branch is the default branch. Can't remove the branch.")
+      } else {
+        try {
+          log(`Removing local branch '${branchName}'.`)
+          await deleteBranch(branchName)
+        } catch (deleteError) {
+          logError(`Could not remove local branch '${branchName}'.`)
+          throw(deleteError)
+        }
       }
     }
     log(`Checking out '${branchName}' from remote.`)
     await checkoutBranch(branchName)
-  } catch(error) {
+  } catch (error) {
     logError("Could not re-checkout branch.")
     throw(error)
   }
@@ -319,9 +452,19 @@ const reCheckoutBranchFromRemote = async branchName => {
 const removeJoukkoBranchFile = async () => {
   const repoRoot = await getPathForGitRoot()
   try {
-    fs.unlinkSync(`${repoRoot}/${JOUKKO_BRANCH_FILE}`)
-  } catch(err) {
-    logError(err)
+    removeFile(repoRoot, JOUKKO_FILE)
+  } catch (error) {
+    logError(error)
+  }
+}
+
+const renameCurrentJoukkoBranch = async (newBranchname) => {
+  try {
+    await renameBranch(newBranchname)
+    await updateJoukkoBranchFileBranch(newBranchname)
+  } catch (error) {
+    logError("Could not rename current joukko branch.")
+    throw(error)
   }
 }
 
@@ -343,15 +486,17 @@ const updateBranch = async (branchName, remoteAlreadyUpdated = false) => {
           try {
             await abortRebase()
             log("Rebase aborted")
-          } catch(rebaseAbortError) {
-            logWarning("Could not abort rebase.") 
+          } catch (rebaseAbortError) {
+            logWarning("Could not abort rebase.")
+            throw(rebaseAbortError)
           }
           const confirmQuestion = `Would you like to remove the current local branch and checkout from remote?`
           const confirmRemoval = await askUserConfirmation(confirmQuestion)
           if (confirmRemoval) {
-            reCheckoutBranchFromRemote()
+            await reCheckoutBranchFromRemote(branchName)
+            log(`Branch '${branchName}' updated.`)
           } else {
-            throw(error)
+            throw("Not removing the local branch or checkout from remote")            
           }
         } else {
           log("Branch update failed.")
@@ -360,29 +505,51 @@ const updateBranch = async (branchName, remoteAlreadyUpdated = false) => {
         }
       }
   } else {
-    log(`Branch '${branchName}' doesn't exist in origin. Not updating.`)
+    log(`Branch '${branchName}' doesn't exist in remote. Not updating.`)
   }
 }
 
+const updateJoukkoBranchFileBranch = async (branchName) => {
+  log("Updating branch name in joukko branch file.")
+  await removeJoukkoBranchFile()
+  await createJoukkoBranchFile(branchName)
+}
+
 module.exports = {
+  abort,
   addJoukkoFileUnderGit,
+  askBranchFromExisting,
+  askOption,
   askUserConfirmation,
   askUserForBranchName,
   askUserInput,
+  changeToBranch,
+  changeToDefaultBranch,
   changeToJoukkoMobBranch,
-  checkIfJoukkoHasAlreadyBeenStarted,
   checkJoukkoFile,
+  checkoutNewLocalBranch,
   checkoutUpToDateBranch,
   createJoukkoBranchFile,
+  getBranchNamesForListing,
+  getBranchNameFromRemotePath,
+  getBranchNamesFromRemotePathBranchList,
   isBranchCurrentBranch,
+  isCurrentBranchTheDefaultBranch,
+  isSelectedOptionValid,
   joukkoFileEmpty,
   joukkoFileFound,
+  joukkoFileHasBranchName,
   preCheckForFinishAndPassOk,
+  preCheckForRenameOk,
   preCheckForTakeOk,
   preCheckForStartOk,
   pushToBranch,
   pushToMobBranch,
+  readJoukkoFileBranch,
   readJoukkoFileContent,
+  reCheckoutBranchFromRemote,
   removeJoukkoBranchFile,
+  renameCurrentJoukkoBranch,
   updateBranch,
+  updateJoukkoBranchFileBranch,
 }
